refactor: remove 'visible' folder from features
