+++ src/features/adaptive.tsx
