+++ src/layout/Header/CatalogMenu/CatalogMenu.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { IsDesktop, IsMobile } from '@/features/visible/adaptive'
+import { IsDesktop, IsMobile } from '@/features/adaptive'
 import MobileCatalogMenu from './MobileCatalogMenu'
 import DesktopCatalogMenu from '@/layout/Header/CatalogMenu/DesktopCatalogMenu'
