+++ src/ui/popups/FilterPopup/FilterPopup.tsx
@@ -5,11 +5,11 @@ import ClosePopupButton from '@/ui/buttons/ClosePopupButton'
 import Button from '@/ui/buttons/Button'
 import { getFilters } from '@/globals/api'
 import { FilterBlock, QuickFilters } from '@/types/FiltersType'
-import { IsDesktop, IsMobile } from '@/features/visible/adaptive'
+import { IsDesktop, IsMobile } from '@/features/adaptive'
 import { FilterItems } from '@/components/FilterItems'
 import ResetFiltersButton from '@/components/QuickFilter/ResetFiltersButton'
 import FilterBlockWrapper from '@/ui/popups/FilterPopup/FilterBlockWrapper'
-import { DynamicPopup, PopupContext } from '@/features/visible/Popup'
+import { DynamicPopup, PopupContext } from '@/features/Popup'
 import FilterTags from '@/components/FilterTags'
 import SortButton from '@/components/QuickFilter/SortButton'
 import { CategoryContext } from '@/layout/catalog/CategoryContext'
